Recycler view only displays one item
Recycler view only displays one item
Changed method signature of IHistorical price from Float to double

+++ app/src/main/java/com/larkspur/stockly/Models/HistoricalPrice.java
@@ -6,19 +6,19 @@ public class HistoricalPrice implements IHistoricalPrice{
 
     private int _historicalPriceId;
     private IStock _stock;
-    private List<Float> _historicalPrice;
+    private List<Double> _historicalPrice;
 
-    public HistoricalPrice(List<Float> historicalPrice){
+    public HistoricalPrice(List<Double> historicalPrice){
         _historicalPrice = historicalPrice;
     }
 
     @Override
-    public Float getPrice() {
+    public Double getPrice() {
         return _historicalPrice.get(_historicalPrice.size()-1);
     }
 
     @Override
-    public List<Float> getHistoricalPrice() {
+    public List<Double> getHistoricalPrice() {
         return _historicalPrice;
     }
