+++ app/src/main/java/com/larkspur/stockly/Models/IStock.java
@@ -10,7 +10,7 @@ public interface IStock {
 
     public Category getCategory();
 
-    public Float getPrice();
+    public Double getPrice();
 
     public IHistoricalPrice getHistoricalPrice();
 }
